+++ packages/zenscript/src/module.ts
@@ -2,8 +2,8 @@ import { type Module, type NameProvider, inject } from 'langium'
 import { type DefaultSharedModuleContext, type LangiumServices, type LangiumSharedServices, type PartialLangiumServices, createDefaultModule, createDefaultSharedModule } from 'langium/lsp'
 import { IntelliZenGeneratedModule, IntelliZenGeneratedSharedModule } from './generated/module'
 import { IntelliZenValidator, registerValidationChecks } from './validator'
-import { ZenScriptScopeProvider } from './scoping/member-scoping'
-import { ZenScriptScopeComputation } from './scoping/precomputed-scoping'
+import { ZenScriptScopeProvider } from './scoping/provider'
+import { ZenScriptScopeComputation } from './scoping/computation'
 import { CustomTokenBuilder } from './lexer/token-builder'
 import { CustomValueConverter } from './lexer/value-converter'
 import { type QualifiedNameProvider, ZenScriptNameProvider } from './name'
