+++ packages/zenscript/src/scoping/provider.ts
